feat: Move list of generations from index to list form
