+++ addon/routes/fd-generation.js
@@ -3,6 +3,9 @@
 */
 
 import Ember from 'ember';
+import { Query } from 'ember-flexberry-data';
+
+const { Builder } = Query;
 
 /**
   Parent route for all forms of generation.
@@ -33,5 +36,36 @@ export default Ember.Route.extend({
         this.transitionTo('fd-generation.log', result.value);
       });
     },
+  },
+
+  /**
+    See [EmberJS API](https://emberjs.com/api/).
+
+    @method model
+  */
+  model() {
+    let store = this.get('store');
+    let modelName = 'fd-generation';
+    let projectionName = 'ListFormView';
+
+    let builder = new Builder(store)
+      .from(modelName)
+      .selectByProjection(projectionName)
+      .where('stage', 'eq', this.get('projectContext').getCurrentStage())
+      .orderBy('startTime desc');
+
+    return Ember.RSVP.hash({
+      generations: store.query(modelName, builder.build()),
+    });
+  },
+
+  /**
+    See [EmberJS API](https://emberjs.com/api/).
+
+    @method afterModel
+  */
+  afterModel({ generations }) {
+    let route = generations.get('length') === 0 ? 'first' : 'list';
+    this.transitionTo(`fd-generation.${route}`);
   },
 });
