--- app/routes/fd-generation/index.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-designer/routes/fd-generation/index';
