+++ app/routes/fd-generation/list.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/routes/fd-generation/list';
